harry

+++ green-library/src/main/java/com/library/service/user/InquiryServiceImpl.java
@@ -24,6 +24,19 @@ public class InquiryServiceImpl implements InquiryService{
 	InquiryRepository inquiryRepository;
 	
 	@Override
+	public boolean checkRentCondition(String userId, String id) {
+		try {
+            String check = inquiryRepository.checkRentCondition(userId, id);
+            if(check.equals("0"))
+            	return false;
+            else
+            	return true;
+        } catch (DataAccessException e) {
+            throw new DatabaseException("Database error occurred while checking user's rent condition with id: " + userId, e);
+        }
+	}
+	
+	@Override
 	public List<UserRentHistoryDTO> getUserRentHistory(String userId) {
 		try {
             return inquiryRepository.getUserRentHistory(userId);
@@ -61,13 +74,14 @@ public class InquiryServiceImpl implements InquiryService{
 	
 	@Override
 	@Transactional(propagation = Propagation.REQUIRED, rollbackFor = Exception.class)
-	public void deleteRentHistory(String id) {
-		System.out.println(id);
+	public void deleteRentHistory(String userId, String id) {
 		try {
-			int rowsAffected = inquiryRepository.deleteRentHistory(id);
-			if (rowsAffected == 0) {
-	            throw new DatabaseException("Failed to delete rent history with id: " + id);
-	        }
+			if(checkRentCondition(userId, id)) {
+				int rowsAffected = inquiryRepository.deleteRentHistory(id);
+				if (rowsAffected == 0) {
+		            throw new DatabaseException("Failed to delete rent history with id: " + id);
+		        }
+			}
         } catch (DataAccessException e) {
             throw new DatabaseException("Database error occurred while removing user's rent history with id: " + id, e);
         }
